Forcing fileuploads to always be an array
When posting a single file to node.js with express, it is being added to
req.files[fieldname] as an object. The changes enforce that the
req.files[fieldname] is always an array containing all submitted files
as objects. Multiple files are automatically properly split up into an
array.

For PHP the same solution is necessary for single files, but for
multiple files we need to change the fieldname by appending square
brackets, so PHP properly recognizes them.

We assume node.js as the default backend, if PHP is used "backendIsPHP"
must be set to true to ensure proper field-naming.